requirements-dbt.txt
dbt-duckdb==1.8.4
